third_party/WebKit/LayoutTests/battery-status/no-gc-with-eventlisteners-expected.txt

Test no garbage collection of battery manager object when listeners are attached.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS batteryInfo.charging is mockBatteryInfo.charging
PASS batteryInfo.chargingTime is mockBatteryInfo.chargingTime
PASS batteryInfo.dischargingTime is mockBatteryInfo.dischargingTime
PASS batteryInfo.level is mockBatteryInfo.level
PASS successfullyParsed is true

TEST COMPLETE
